+++ app/controllers/recording_studio_terms_and_conditions/admin/base_controller.rb
@@ -3,6 +3,9 @@
 module RecordingStudioTermsAndConditions
   module Admin
     class BaseController < RecordingStudioTermsAndConditions::ApplicationController
+      include ::Pagy::Backend
+      include TablePage
+
       before_action :authenticate_user!, raise: false
       before_action :require_admin_access!
 
@@ -45,7 +48,7 @@ module RecordingStudioTermsAndConditions
 
       def terms_recording
         @terms_recording ||= RecordingStudio::Recording.find_by!(
-          id: params[:id],
+          id: params[:term_id] || params[:id],
           recordable_type: Terms.name,
           trashed_at: nil
         )